- Extracts the TitleEditor from header and adds it to a new file (title-editor.js). - Minor refactoring to use the proper 'uncontrolled input approach'.
